Move RolloutSavingWrapper to algorithm utils
